Added render selection

+++ icon-maker-addon/renderer.py
@@ -12,9 +12,10 @@ def createOutlineObject(object, thickness):
     # Copy model
     utils.selectObject(object)
     outlineObject = object.copy()
-    utils.setData(outlineObject)
     outlineObject.data = object.data.copy()
+    utils.setData(outlineObject)
     outlineObject.name = object.name + "_outline"
+    outlineObject.data.name = object.name + "_outline"
     outlineObject.data.materials.clear()
     scene.collection.objects.link(outlineObject)
     
@@ -55,6 +56,7 @@ def createShadowObject(object):
     
     shadowObject = object.copy()
     shadowObject.data = object.data.copy()
+    utils.setData(shadowObject)
     shadowObject.name = object.name + "_shadow"
     shadowObject.data.materials.clear()
     
@@ -69,12 +71,12 @@ def createShadowObject(object):
     
     return shadowPlane, shadowObject
 
-def renderFrames(name):
+def renderFrame(output):
     
     scene = bpy.context.scene
 
     # Settings
-    scene.render.filepath = scene.icomake_props.output_dir + name + ".tga"
+    scene.render.filepath = output
     scene.render.resolution_x = 512
     scene.render.resolution_y = 512
     scene.render.image_settings.file_format='TARGA'
@@ -85,33 +87,8 @@ def renderFrames(name):
 
 # Call Functions
 
-def makeIcon(pgroup):
+def makeIcon(object, pos = "FLOOR", outline = 0, render_output = "//"):
     scene = bpy.context.scene
-    modelpth = os.path.join(bpy.path.abspath("//") + pgroup.path, pgroup.name)
-    type = pgroup.position
-    
-    if "smd" in os.path.splitext(modelpth)[1]:
-        object = utils.importSourceModel(modelpth)
-    elif "obj" in os.path.splitext(modelpth)[1]:
-        object = utils.importObj(modelpth)
-        object.rotation_euler = ([radians(a) for a in (0.0, 0.0, 90.0)])
-    
-    utils.setData(object)
-
-    for material_slot in object.material_slots:
-        material = material_slot.material
-        utils.setData(material)
-        
-        # Get image for material
-        imagepth = os.path.join(os.path.dirname(modelpth), material.name + ".tga")
-        if os.path.exists(imagepth):
-            image = bpy.data.images.load(filepath=imagepth)
-        else:
-            image = bpy.data.images.new(name=material.name + ".tga", width=16, height=16)
-            image.generated_color = (1, 0, 0.5, 1)
-        utils.setData(image)
-
-    node_utils.nodesMatModel(material, image)
 
     # camera and object placement
     camera_data = bpy.data.cameras.new("[ICOMAKE] Camera")
@@ -119,11 +96,11 @@ def makeIcon(pgroup):
     utils.setData(camera)
     
     # Place cameras
-    if type == "FLOOR":
+    if pos == "FLOOR":
         camera.rotation_euler = ([radians(a) for a in (60.0, 0.0, 330.0)])
-    elif type == "WALL":
+    elif pos == "WALL":
         camera.rotation_euler = ([radians(a) for a in (330.0, 330.0, 180.0)])
-    elif type == "CEIL":
+    elif pos == "CEIL":
         camera.rotation_euler = ([radians(a) for a in (60.0, 180.0, 390.0)])
     
     camera.data.type = "ORTHO"
@@ -148,11 +125,12 @@ def makeIcon(pgroup):
     tempCol.objects.link(camera)
     
     # Create extra objects for rendering
-    outlineObject = createOutlineObject(object, pgroup.outline)
-    scene.collection.objects.unlink(outlineObject)
-    tempColOut.objects.link(outlineObject)
+    if not outline == 0:
+        outlineObject = createOutlineObject(object, outline)
+        scene.collection.objects.unlink(outlineObject)
+        tempColOut.objects.link(outlineObject)
     
-    if not type == "CEIL":
+    if not pos == "CEIL":
         shadowPlane, shadowObject = createShadowObject(object)
         bpy.context.collection.objects.unlink(shadowPlane)
         tempColSdw.objects.link(shadowPlane)
@@ -195,24 +173,71 @@ def makeIcon(pgroup):
     # Compositing
     node_utils.nodesCompositing(objectLayer, outlineLayer, shadowLayer)
 
-    renderFrames(os.path.splitext(pgroup.name)[0])
+    renderFrame(render_output)
 
 def setupScene():
     # Sun
-    sun_data = bpy.data.lights.new("[ICOMAKE] Sun")
+    scene = bpy.context.scene
+
+    sun_data = bpy.data.lights.new("[ICOMAKE] Sun", type="SUN")
     sun = bpy.data.objects.new("[ICOMAKE] Sun", sun_data)
     utils.setData(sun, "icomake_scenedata")
 
+    sun.data.color = (0.603827, 0.603827, 0.603827)
+    sun.data.energy = 1
+    sun.data.diffuse_factor = 1
+    sun.data.specular_factor = 1
+    sun.data.volume_factor = 1
+    sun.data.angle = 0
+
+    sun.data.shadow_buffer_bias = 0.002
+
     sun.rotation_euler = ([radians(a) for a in (45.0, 0.0, 20.0)])
 
-class IM_MassRender(bpy.types.Operator):
+    if not scene.collection.objects.get(sun.name):
+        scene.collection.objects.link(sun)
+
+def RenderMass(pgroup):
+    scene = bpy.context.scene
+    
+    modelpth = os.path.join(bpy.path.abspath("//") + pgroup.path, pgroup.name)
+    pos = pgroup.position
+
+    if "smd" in os.path.splitext(modelpth)[1]:
+        object = utils.importSourceModel(modelpth)
+    elif "obj" in os.path.splitext(modelpth)[1]:
+        object = utils.importObj(modelpth)
+        object.rotation_euler = ([radians(a) for a in (0.0, 0.0, 90.0)])
+
+    utils.setData(object)
+
+    for material_slot in object.material_slots:
+        material = material_slot.material
+        utils.setData(material)
+        
+        # Get image for material
+        imagepth = os.path.join(os.path.dirname(modelpth), material.name + ".tga")
+        if os.path.exists(imagepth):
+            image = bpy.data.images.load(filepath=imagepth)
+        else:
+            image = bpy.data.images.new(name=material.name + "_temp", width=16, height=16)
+            image.generated_color = (1, 0, 0.5, 1)
+        utils.setData(image)
+
+        node_utils.nodesMatModel(material, image)
+
+    render_output = scene.icomake_props.rendermass_output + os.path.splitext(pgroup.name)[0] + ".tga"
+
+    makeIcon(object, pos, pgroup.outline, render_output)
+
+class IM_RenderMass(bpy.types.Operator):
     """Render models and export icons"""
-    bl_idname = "icomake.massrender"
+    bl_idname = "icomake.rendermass"
     bl_label = "Mass Render Icons"
     
     @classmethod
     def poll(cls, context):
-        return len(context.scene.icomake_imports) > 0 and context.scene.icomake_props.output_dir != ""
+        return len(context.scene.icomake_rendermass_imports) > 0 and context.scene.icomake_props.rendermass_output != ""
     
     def execute(self, context):
         scene = context.scene
@@ -220,13 +245,85 @@ class IM_MassRender(bpy.types.Operator):
         utils.cleanUpData("icomake_scenedata")
         utils.cleanUpData("icomake_tempdata")
 
-        setupScene
+        setupScene()
 
-        for pgroup in scene.icomake_imports:
-            makeIcon(pgroup)
+        for pgroup in scene.icomake_rendermass_imports:
+            RenderMass(pgroup)
             utils.cleanUpData("icomake_tempdata")
             
         utils.cleanUpData("icomake_scenedata")
-            
         
+        return {'FINISHED'}
+
+class IM_RenderSelected(bpy.types.Operator):
+    """Render model and export icons"""
+    bl_idname = "icomake.renderselected"
+    bl_label = "Render Selected Object"
+    
+    @classmethod
+    def poll(cls, context):
+        return context.active_object is not None and len(context.selected_objects) > 0 and context.scene.icomake_props.renderselected_output != ""
+    
+    def execute(self, context):
+        scene = context.scene
+
+        print("Active Object = " + context.active_object.name)
+        object = context.active_object
+
+        utils.cleanUpData("icomake_scenedata")
+        utils.cleanUpData("icomake_tempdata")
+
+        setupScene()
+
+        render_output = scene.icomake_props.renderselected_output + os.path.splitext(object.name)[0] + ".tga"
+        position = scene.icomake_props.renderselected_position
+        outline = scene.icomake_props.renderselected_outline
+        makeIcon(object, position, outline, render_output)
+
+        utils.cleanUpData("icomake_scenedata")
+        utils.cleanUpData("icomake_tempdata")
+
+        if not scene.collection.objects.get(object.name):
+            scene.collection.objects.link(object)
+        
+        return {'FINISHED'}
+
+class IM_MaterialSelected(bpy.types.Operator):
+    """Apply PM Shader to selected"""
+    bl_idname = "icomake.materialselected"
+    bl_label = "Apply PM Shader to Selected"
+    
+    @classmethod
+    def poll(cls, context):
+        return context.active_object is not None and len(context.selected_objects) > 0
+    
+    def execute(self, context):
+        scene = context.scene
+
+        object = context.active_object
+
+        for material_slot in object.material_slots:
+            material = material_slot.material
+            
+            # Make temp image for material
+            image = bpy.data.images.new(name=material.name + "_temp", width=16, height=16)
+            image.generated_color = (1, 0, 0.5, 1)
+
+            node_utils.nodesMatModel(material, image)
+        
+        return {'FINISHED'}
+
+class IM_CleanUp(bpy.types.Operator):
+    """Apply PM Shader to selected"""
+    bl_idname = "icomake.cleanup"
+    bl_label = "Clean Up ICOMAKE Data"
+    
+    #@classmethod
+    #def poll(cls, context):
+    #    return True
+    
+    def execute(self, context):
+        utils.cleanUpData("icomake_scenedata")
+        utils.cleanUpData("icomake_tempdata")
+                
         return {'FINISHED'}
